cherry-pick 50fc20dc940c87ffd62be0650b935324e164b59d
Fix wallpaper not loading (leaving the background black)
We need to apply the wallpaper anchors after the lockscreen theme is loaded, otherwise they won't be set properly and we end up with a black background.

This issue was triggered by https://invent.kde.org/plasma/kscreenlocker/-/merge_requests/84

The reason why this wasn't noticed on Plasma Desktop, was that its lockscreen theme draws a FastBlur above the wallpaper in the background which will still correctly display the wallpaper item.

I had forgotten about that I used that as a workaround in Plasma Mobile for months... oops